Restructure my project code.
